+++ packages/sdk-client/tests/api/api-client-options-helper.test.ts
@@ -55,6 +55,26 @@ describe('API Client Options helper', () => {
       expect(apiClientOptions.requestPlugins?.length).toBe(1);
       expect(apiClientOptions.requestPlugins?.[0]).toBeInstanceOf(Oauth2TokenRequest);
       expect(apiClientOptions.responsePlugins).toBeUndefined();
+      expect(apiClientOptions.retryPolicy).toBe('DEFAULT');
+      expect(apiClientOptions.maxRetryCount).toBe(3);
+      expect(apiClientOptions.exponentialBackoff).toBe(4);
+    });
+
+    it('should propagate custom retry policy settings', () => {
+      const params: SinchClientParameters = {
+        projectId: 'PROJECT_ID',
+        keyId: 'KEY_ID',
+        keySecret: 'KEY_SECRET',
+        retryPolicy: 'NONE',
+        maxRetryCount: 1,
+        exponentialBackoff: 2,
+      };
+
+      const apiClientOptions = buildOAuth2ApiClientOptions(params, 'foo');
+
+      expect(apiClientOptions.retryPolicy).toBe('NONE');
+      expect(apiClientOptions.maxRetryCount).toBe(1);
+      expect(apiClientOptions.exponentialBackoff).toBe(2);
     });
 
     it('should build some API client options with additional plugins', () => {